requirements.txt
# VTU Results Automation System - Requirements

# Flask web framework
flask==2.3.3
flask-socketio==5.3.6

# Browser automation
playwright==1.40.0

# AI/ML libraries for captcha solving
# google-generativeai==0.3.2  # REMOVED - Using local model instead
pillow==10.1.0

# Environment management
python-dotenv==1.0.0

# Additional utilities
pathlib2==2.3.7
requests==2.31.0

# PyTorch for local CAPTCHA model
torch==2.1.0
torchvision==0.16.0

# Development dependencies (optional)
# pytest==7.4.3
# pytest-flask==1.3.0